requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
python-dotenv>=1.0.1
langchain-core>=0.1.0
langchain>=0.1.9
langchain-openai>=0.0.8
langchain-chroma>=0.2.4
chromadb>=0.4.22
openai>=1.12.0
pydantic>=2.6.1
python-multipart>=0.0.9 